Rename examples Tailwind config files to kebab-case
