+++ packages/examples/src/config/tailwind/index.ts
@@ -1,2 +1,2 @@
-export * from './autofillOverride';
-export * from './stateLayer';
+export * from './autofill-override';
+export * from './state-layer';
